+++ source/Karamem0.SPClientCore/Commands/RemoveGroupMemberCommand.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2023 karamem0
+// Copyright (c) 2018-2024 karamem0
 //
 // This software is released under the MIT License.
 //
@@ -15,29 +15,29 @@ using System.Linq;
 using System.Management.Automation;
 using System.Text;
 
-namespace Karamem0.SharePoint.PowerShell.Commands
+namespace Karamem0.SharePoint.PowerShell.Commands;
+
+[Cmdlet(VerbsCommon.Remove, "KshGroupMember", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.High)]
+[OutputType((Type[])null)]
+public class RemoveGroupMemberCommand : ClientObjectCmdlet<IGroupMemberService>
 {
 
-    [Cmdlet("Remove", "KshGroupMember")]
-    [OutputType(typeof(void))]
-    public class RemoveGroupMemberCommand : ClientObjectCmdlet<IGroupMemberService>
+    public RemoveGroupMemberCommand()
     {
+    }
 
-        public RemoveGroupMemberCommand()
-        {
-        }
+    [Parameter(Mandatory = true, Position = 0)]
+    public Group Group { get; private set; }
 
-        [Parameter(Mandatory = true, Position = 0)]
-        public Group Group { get; private set; }
+    [Parameter(Mandatory = true, Position = 1)]
+    public User Member { get; private set; }
 
-        [Parameter(Mandatory = true, Position = 1)]
-        public User Member { get; private set; }
-
-        protected override void ProcessRecordCore()
+    protected override void ProcessRecordCore()
+    {
+        if (this.ShouldProcess(this.Member.Title, VerbsCommon.Remove))
         {
             this.Service.RemoveObject(this.Group, this.Member);
         }
-
     }
 
 }
