+++ source/Karamem0.SPClientCore/Commands/RemoveSiteFeatureCommand.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2023 karamem0
+// Copyright (c) 2018-2024 karamem0
 //
 // This software is released under the MIT License.
 //
@@ -15,29 +15,29 @@ using System.Linq;
 using System.Management.Automation;
 using System.Text;
 
-namespace Karamem0.SharePoint.PowerShell.Commands
+namespace Karamem0.SharePoint.PowerShell.Commands;
+
+[Cmdlet(VerbsCommon.Remove, "KshSiteFeature", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.High)]
+[OutputType((Type[])null)]
+public class RemoveSiteFeatureCommand : ClientObjectCmdlet<ISiteFeatureService>
 {
 
-    [Cmdlet("Remove", "KshSiteFeature")]
-    [OutputType(typeof(void))]
-    public class RemoveSiteFeatureCommand : ClientObjectCmdlet<ISiteFeatureService>
+    public RemoveSiteFeatureCommand()
     {
+    }
 
-        public RemoveSiteFeatureCommand()
-        {
-        }
+    [Parameter(Mandatory = true, Position = 0, ValueFromPipeline = true)]
+    public Feature Identity { get; private set; }
 
-        [Parameter(Mandatory = true, Position = 0, ValueFromPipeline = true)]
-        public Feature Identity { get; private set; }
+    [Parameter(Mandatory = false)]
+    public bool Force { get; private set; }
 
-        [Parameter(Mandatory = false)]
-        public bool Force { get; private set; }
-
-        protected override void ProcessRecordCore()
+    protected override void ProcessRecordCore()
+    {
+        if (this.ShouldProcess(this.Identity.DisplayName, VerbsCommon.Remove))
         {
             this.Service.RemoveObject(this.Identity.DefinitionId, this.Force);
         }
-
     }
 
 }
